+++ R/slidy_presentation.R
@@ -2,12 +2,12 @@
 #'
 #' Format for converting from R Markdown to a slidy presentation.
 #'
-#' See the \href{https://bookdown.org/yihui/rmarkdown/slidy-presentation.html}{online
+#' See the \href{https://yihui.org/rmarkdown/slidy-presentation.html}{online
 #' documentation} for additional details on using the \code{slidy_presentation}
 #' format.
 #'
 #' For more information on markdown syntax for presentations see the
-#' \href{https://pandoc.org/README.html}{pandoc online documentation}.
+#' \href{https://pandoc.org/MANUAL.html}{pandoc online documentation}.
 #' @inheritParams pdf_document
 #' @inheritParams html_document
 #' @inheritParams beamer_presentation