+++ NEWS.md
@@ -1,6 +1,8 @@
 rmarkdown 2.31
 ================================================================================
 
+- Base64 encoded images work with non-HTML output formats now (thanks, @averissimo, #2604).
+
 - Use the argument `--syntax-highlighting` for Pandoc >= 3.8 since `--highlight-style` has been deprecated (thanks, @ywwry66, #2602).